+++ MessageApp/src/main/java/com/example/accessingdatajpa/Models/Repository/QueueRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface QueueRepository extends JpaRepository<Queue, Integer> {
+public interface QueueRepository extends JpaRepository<Queue, Long> {
     Optional<Queue> findByName(String name);
 }
